chore!: deprecate numpy < 2.0 and python < 3.11 (#1163)
This PR deprecates support for Python <3.11 and NumPy <2.0 to align with updated dependencies (shap, xgboost, skops). The changes include:

* Updating minimum Python version to 3.11 and adding support for Python 3.13
* Updating NumPy to >=2.3.0, hail to >=0.2.136, and shap to >=0.50.0
* Removing typing_extensions dependency (now available in Python 3.11+ standard library)
* Replacing deprecated NumPy APIs (np.Infinity → np.inf)
* Replacing PySpark's assertDataFrameEqual with direct list comparisons
* Updating doctests to handle numpy 2.0's changed string representation (np.float64(x))
* Migrating from pyspark.pandas API to standard pandas for reading parquet files

---------

+++ src/gentropy/method/carma.py
@@ -167,7 +167,7 @@ class CARMA:
             >>> p_S = 3
             >>> det_S = 0.1
             >>> np.round(CARMA._ind_normal_sigma_fixed_marginal_fun_indi(zSigmaz_S, tau, p_S, det_S),decimals=5)
-            10.18849
+            np.float64(10.18849)
         """
         return p_S / 2.0 * np.log(tau) - 0.5 * np.log(det_S) + zSigmaz_S / 2.0
 
@@ -198,7 +198,7 @@ class CARMA:
             >>> tau = 1
             >>> p_S = 2
             >>> np.round(CARMA._ind_Normal_fixed_sigma_marginal_external(index_vec_input, Sigma, z, tau, p_S),decimals=5)
-            43.60579
+            np.float64(43.60579)
         """
         index_vec = index_vec_input - 1
         Sigma_S = Sigma[np.ix_(index_vec, index_vec)]
@@ -243,7 +243,7 @@ class CARMA:
             >>> tau = 1 / 0.05**2
             >>> p_S = 3
             >>> np.round(CARMA._outlier_ind_Normal_marginal_external(index_vec_input, Sigma, z, tau, p_S),decimals=5)
-            -8.8497
+            np.float64(-8.8497)
         """
         index_vec = index_vec_input - 1
 
@@ -488,7 +488,7 @@ class CARMA:
         >>> outlier_tau = 1 / 0.05**2
         >>> outlier_likelihood = CARMA._outlier_ind_Normal_marginal_external
         >>> np.round(CARMA._ridge_fun(x, Sigma, modi_ld_S, test_S, z, outlier_tau, outlier_likelihood),decimals=5)
-        6.01486
+        np.float64(6.01486)
         """
         temp_Sigma = Sigma.copy()
         temp_ld_S = x * modi_ld_S + (1 - x) * np.eye(len(modi_ld_S))
@@ -518,7 +518,7 @@ class CARMA:
         >>> lambda_val = 1
         >>> p = 4
         >>> np.round(CARMA._prior_dist(t, lambda_val, p),decimals=5)
-        -3.17805
+        np.float64(-3.17805)
         """
         index_array = t.split(",")
         dim_model = len(index_array)
